+++ cmd/podcertcontroller/main.go
@@ -30,11 +30,11 @@ import (
 	"path/filepath"
 	"syscall"
 
+	"github.com/agent-substrate/substrate/cmd/podcertcontroller/internal/podidentitysigner"
+	"github.com/agent-substrate/substrate/cmd/podcertcontroller/internal/rendezvous"
+	"github.com/agent-substrate/substrate/cmd/podcertcontroller/internal/servicednssigner"
+	"github.com/agent-substrate/substrate/cmd/podcertcontroller/internal/signercontroller"
 	"github.com/agent-substrate/substrate/internal/localca"
-	"github.com/agent-substrate/substrate/internal/podidentitysigner"
-	"github.com/agent-substrate/substrate/internal/rendezvous"
-	"github.com/agent-substrate/substrate/internal/servicednssigner"
-	"github.com/agent-substrate/substrate/internal/signercontroller"
 	"github.com/agent-substrate/substrate/internal/version"
 	"github.com/spf13/pflag"
 	"k8s.io/apimachinery/pkg/types"
